Refactored : Extract class
